+++ src/views/Home.vue
@@ -12,10 +12,12 @@
             span(v-html='announcement.description')
         br
         a.btn.btn-warning(href='https://www.youtube.com/watch?v=fxuAnP2GF0I&feature=youtu.be', target='_blank') Watch our video!
-    // .col-md-8
-    //   div.carousel(uib-carousel='', active='1', interval='7000', no-wrap='false')
-    //     div(uib-slide='', ng-repeat='slide in $ctrl.slides', active='active', index='$index')
-    //       img.img-responsive(ng-src='{{slide.src}}', alt='Person practicing at one of our Pittsburgh yoga schools')
+    .col-md-8
+      b-carousel(id='carouselFade', interval='7000', fade='', indicators='')
+        b-carousel-slide(img-src='/img/home1.jpg', img-alt='Woman in handstand at one of our Pittsburgh yoga schools')
+        b-carousel-slide(img-src='/img/home2.jpg', img-alt='Woman in lotus at one of our Pittsburgh yoga schools')
+        b-carousel-slide(img-src='/img/home3.jpg', img-alt='Woman stretching at one of our Pittsburgh yoga schools')
+        b-carousel-slide(img-src='/img/home4.jpg', img-alt='Man in lotus at one of our Pittsburgh yoga schools')
   .row
     h1.col-md-6 Frequently-asked questions
   .row
@@ -32,11 +34,16 @@ import { Component, Vue } from 'vue-property-decorator';
 export default class Home extends Vue {
   private mounted() {
     this.announcementsFetch();
+    this.carouselFetch();
     this.faqsFetch();
   }
 
   get announcementSections(): any[] {
     return this.$store.state.announcementSections;
+  }
+
+  get carousel(): string[] {
+    return this.$store.state.carousel;
   }
 
   get faqs(): any[] {
@@ -45,6 +52,10 @@ export default class Home extends Vue {
 
   private announcementsFetch() {
     this.$store.dispatch('announcementsFetch');
+  }
+
+    private carouselFetch() {
+    this.$store.dispatch('carouselFetch');
   }
 
   private faqsFetch() {
@@ -82,14 +93,12 @@ export default class Home extends Vue {
 
 h1 {
   font-size: 30px;
+  margin-top: 10px;
+  margin-bottom: 0;
 }
 
 h2 {
   font-size: 18px;
-}
-
-.carousel-control {
-  display: none;
 }
 
 $column-gap: 3em;